Change maxMessages type from int to short

+++ src/XgpLib.SyncService.Api/Controllers/MessageBrokerController.cs
@@ -56,7 +56,7 @@ public class MessageBrokerController(
     [HttpGet("receive/{topic}")]
     public async Task<ActionResult<ReceiveMessagesResponse>> ReceiveMessages(
         [FromRoute] string topic,
-        [FromQuery] int maxMessages = 10,
+        [FromQuery] short maxMessages = 10,
         CancellationToken cancellationToken = default)
     {
         if (string.IsNullOrWhiteSpace(topic))
